% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5c_Volume_of_peat_drained.R
\name{area_drained}
\alias{area_drained}
\title{area_drained}
\usage{
area_drained(drain_ext, l, w)
}
\arguments{
\item{drain_ext}{average extent of drainage around drainage features}

\item{l}{length of drainage feature}

\item{w}{width of drainage feature}
}
\value{
Area of drained peat (assuming rectangular drainage profile)
}
\description{
area_drained
}
