% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5c_Volume_of_peat_drained.R
\name{AV_peat_drained}
\alias{AV_peat_drained}
\title{AV_peat_drained}
\usage{
AV_peat_drained(core.dat, construct.dat)
}
\arguments{
\item{core.dat}{UI data}

\item{construct.dat}{UI construction data}
}
\value{
AV_indirect
}
\description{
AV_peat_drained
}
