% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_Volume_of_peat_removed.R
\name{AV_peat_removed}
\alias{AV_peat_removed}
\title{AV_peat_removed}
\usage{
AV_peat_removed(core.dat, construct.dat)
}
\arguments{
\item{core.dat}{UI data}

\item{construct.dat}{UI construction data}
}
\value{
AV_direct
}
\description{
AV_peat_removed
}
