% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7b_C_seq_in_soil_under_trees.R
\docType{data}
\name{C.seq.in.soil.under.trees}
\alias{C.seq.in.soil.under.trees}
\title{C.seq.in.soil.under.trees}
\format{
An object of class \code{data.frame} with 1 rows and 2 columns.
}
\usage{
C.seq.in.soil.under.trees
}
\value{
C.seq.in.soil.under.trees
}
\description{
C.seq.in.soil.under.trees
}
\keyword{datasets}
