% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5d_CO2_loss_from_drained_peat.R
\name{CO2_loss_drained}
\alias{CO2_loss_drained}
\title{CO2_loss_drained}
\usage{
CO2_loss_drained(
  pC_dry_peat,
  BD_dry_soil,
  restore_hydr_in,
  A_indirect,
  V_indirect,
  t_wf,
  t_restore
)
}
\arguments{
\item{pC_dry_peat}{Carbon content of dry peat}

\item{BD_dry_soil}{Dry soil bulk density}

\item{restore_hydr_in}{Select hydrology restoration}

\item{A_indirect}{Area peat drained}

\item{V_indirect}{Volume peat drained}

\item{t_wf}{Windfarm life time}

\item{t_restore}{Time to restoration after decomissioning}

\item{E_tot}{Total emissions of undrained per per unit area}

\item{CO2_C}{Molecular weight ratio C to CO2}

\item{pCO2_lost}{percent Carbon lost as CO2}
}
\value{
Hypothetical carbon fixation of drained/removed peat
}
\description{
CO2_loss_drained
}
