% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Loss_of_soil_CO2.R
\name{CO2_loss_from_soil}
\alias{CO2_loss_from_soil}
\title{CO2_loss_from_soil}
\usage{
CO2_loss_from_soil(L_direct, L_indirect)
}
\arguments{
\item{L_direct}{Carbon content of dry peat}

\item{L_indirect}{Dry soil bulk density}
}
\value{
Total CO2 loss from removed and drained
}
\description{
CO2_loss_from_soil
}
