% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_CO2_gain_site_improvement.R
\name{CO_2_gain_site_improve}
\alias{CO_2_gain_site_improve}
\title{CO_2_gain_site_improve}
\usage{
CO_2_gain_site_improve(core.dat)
}
\arguments{
\item{core.dat}{UI data}
}
\value{
L_improvement
}
\description{
CO_2_gain_site_improve
}
