% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_CO2_gain_site_improvement.R
\name{dat_improvement}
\alias{dat_improvement}
\title{dat_improvement}
\usage{
dat_improvement(core.dat)
}
\arguments{
\item{core.dat}{UI data}
}
\value{
d/a/t_feature
}
\description{
dat_improvement
}
