% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5c_Volume_of_peat_drained.R
\name{drainage}
\alias{drainage}
\title{drainage}
\usage{
drainage(drain_ext, pit_dims, fh_dims, at_dims, ct_dims, add_dims)
}
\arguments{
\item{drain_ext}{average extent of drainage around drainage features}

\item{pit_dims}{borrow pit dimensions (list)}

\item{fh_dims}{foundations + hardstanding dimensions}

\item{at_dims}{access track dimensions (list)}

\item{ct_dims}{cable trench  dimensions (list)}

\item{add_dims}{additional excavation dimensions (list)}
}
\value{
Area/Volume of peat drained
}
\description{
drainage
}
