% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5c_Volume_of_peat_drained.R
\name{drainage_add}
\alias{drainage_add}
\title{drainage_add}
\usage{
drainage_add(drain_ext, add_dims)
}
\arguments{
\item{drain_ext}{average extent of drainage around drainage features}

\item{add_dims}{additional excavation dimensions (list)}
}
\value{
Area/Volume of peat drained due to additional excavations
}
\description{
drainage_add
}
