% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5c_Volume_of_peat_drained.R
\name{drainage_at}
\alias{drainage_at}
\title{drainage_at}
\usage{
drainage_at(drain_ext, at_dims)
}
\arguments{
\item{drain_ext}{average extent of drainage around drainage features}

\item{at_dims}{access track dimensions (list)}
}
\value{
Area/Volume of peat drained due to access tracks
}
\description{
drainage_at
}
