% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5c_Volume_of_peat_drained.R
\name{drainage_bp}
\alias{drainage_bp}
\title{drainage_bp}
\usage{
drainage_bp(drain_ext, pit_dims)
}
\arguments{
\item{drain_ext}{average extent of drainage around drainage features}

\item{pit_dims}{number, length, width and depth of borrow pits (list)}
}
\value{
Area/Volume of peat drained due to borrow pits
}
\description{
drainage_bp
}
