% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5c_Volume_of_peat_drained.R
\name{drainage_ct}
\alias{drainage_ct}
\title{drainage_ct}
\usage{
drainage_ct(drain_ext, ct_dims)
}
\arguments{
\item{drain_ext}{average extent of drainage around drainage features}

\item{ct_dims}{cable trench dimensions (list)}
}
\value{
Area/Volume of peat drained due to cable trenches
}
\description{
drainage_ct
}
