% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5c_Volume_of_peat_drained.R
\name{drainage_fh}
\alias{drainage_fh}
\title{drainage_fh}
\usage{
drainage_fh(drain_ext, fh_dims)
}
\arguments{
\item{drain_ext}{average extent of drainage around drainage features}

\item{fh_dims}{foundations + hardstanding dimensions}
}
\value{
Area/Volume of peat drained due to foundations + hardstanding
}
\description{
drainage_fh
}
