% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5e_Emission_rates_from_soils.R
\name{Emissions_rates_soils}
\alias{Emissions_rates_soils}
\title{Emissions_rates_soils}
\usage{
Emissions_rates_soils(
  em_factor_meth_in,
  peat_type,
  A_indirect,
  V_indirect,
  T_air,
  d_wt,
  CO2_C = 3.667,
  CH4_CO2 = 30.67
)
}
\arguments{
\item{em_factor_meth_in}{Select IPCC default or ECOSSE model}

\item{peat_type}{Select acid bog or fen}

\item{A_indirect}{Area peat drained}

\item{V_indirect}{Volume peat drained}

\item{T_air}{Average air temperature}

\item{d_wt}{Average water table depth (undrained)}

\item{CO2_C}{Molecular weight ratio C to CO2}

\item{CH4_CO2}{CH4 to CO2 conversion factor}
}
\value{
Emissions rates from drained and undrained peatland
}
\description{
Emissions_rates_soils
}
