% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7ii_Forestry_CO2_loss_detail.R
\name{Forestry_CO2_loss_detail}
\alias{Forestry_CO2_loss_detail}
\title{Forestry_CO2_loss_detail}
\usage{
Forestry_CO2_loss_detail(core.dat)
}
\arguments{
\item{core.dat}{UI data}
}
\value{
Estimated lifetime loss of carbon sequestration
}
\description{
Forestry_CO2_loss_detail
}
