% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7i_Forestry_CO2_loss_simple.R
\name{Forestry_CO2_loss_simple}
\alias{Forestry_CO2_loss_simple}
\title{Forestry_CO2_loss_simple}
\usage{
Forestry_CO2_loss_simple(core.dat)
}
\arguments{
\item{core.dat}{UI data}
}
\value{
Estimated lifetime loss of carbon sequestration
}
\description{
Forestry_CO2_loss_simple
}
