% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getConstructionInputData}
\alias{getConstructionInputData}
\title{getConstructionInputData}
\usage{
getConstructionInputData(path, df_struct)
}
\arguments{
\item{path}{path to UI}

\item{df_struct}{dataframe summarising the data structure to output}
}
\value{
Construction Input Data
}
\description{
getConstructionInputData
}
