% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getCoreInputData}
\alias{getCoreInputData}
\title{getCoreInputData}
\usage{
getCoreInputData(path, df_struct)
}
\arguments{
\item{path}{path to UI}

\item{df_struct}{dataframe summarising the data structure to output}
}
\value{
Core Input Data
}
\description{
getCoreInputData
}
