% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7c_Average_stand_data.R
\name{growth_yield_tab}
\alias{growth_yield_tab}
\title{growth_yield_tab}
\usage{
growth_yield_tab(t, soil_type, species = 2)
}
\arguments{
\item{t}{stand age}

\item{soil_type}{soil type (deep peat or peaty gley)}

\item{species}{1: scots pine; 2: sitka spruce}
}
\value{
C_seq_soil
}
\description{
growth_yield_tab
}
