% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_CO2_loss_due_to_turbine_life.R
\name{Lifetime_emissions}
\alias{Lifetime_emissions}
\title{Lifetime_emissions}
\usage{
Lifetime_emissions(core.dat, construct.dat, E_concrete = 0.316)
}
\arguments{
\item{core.dat}{UI data}

\item{construct.dat}{UI construction data}

\item{E_cement}{emissions per unit volume cement}
}
\value{
Windfarm lifetime emissions
}
\description{
Lifetime_emissions
}
