% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_CO2_loss_due_to_turbine_life.R
\name{Lifetime_emissions0}
\alias{Lifetime_emissions0}
\title{Lifetime_emissions0}
\usage{
Lifetime_emissions0(
  c_turb,
  n_turb,
  coefs,
  V_concrete,
  c_turb_thresh = 1,
  E_concrete = 0.316
)
}
\arguments{
\item{c_turb}{maximum tubine power capacity}

\item{n_turb}{number of turbines}

\item{coefs}{regression coefficeints}

\item{V_concrete}{total volume of cement used}

\item{c_turb_thresh}{threshold turbine power capacity for regression models}

\item{E_cement}{emissions per unit volume cement}
}
\value{
Windfarm lifetime emissions
}
\description{
Lifetime_emissions0
}
