% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Loss_of_CO2_fixing_pot.R
\name{Loss_of_CO2_fix_pot}
\alias{Loss_of_CO2_fix_pot}
\title{Loss_of_CO2_fix_pot}
\usage{
Loss_of_CO2_fix_pot(core.dat, AV_direct, AV_indirect, CO2_C = 3.667)
}
\arguments{
\item{CO2_C}{Molecular weight ratio C to CO2}

\item{A_direct}{Area of removed peat}

\item{A_indirect}{Area of drained peat}

\item{G_bog}{Peatland carbon accumulation rate}

\item{t_wf}{Windfarm life time}

\item{t_restore}{Time to restoration after decomissioning}
}
\value{
L_fix
}
\description{
Loss_of_CO2_fix_pot
}
