% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_Volume_of_peat_removed.R
\name{removal}
\alias{removal}
\title{removal}
\usage{
removal(pit_dims, f_dims, h_dims, at_dims, add_dims)
}
\arguments{
\item{pit_dims}{borrow pit dimensions (list)}

\item{f_dims}{foundation dimensions (list)}

\item{h_dims}{hardstand dimensions (list)}

\item{at_dims}{access track dimensions (list)}

\item{add_dims}{additional excavation dimensions (list)}
}
\value{
Area/Volume of peat removed
}
\description{
removal
}
