% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_Volume_of_peat_removed.R
\name{removal_at}
\alias{removal_at}
\title{removal_at}
\usage{
removal_at(at_dims)
}
\arguments{
\item{at_dims}{access track dimensions (list)}
}
\value{
Area/Volume of peat removed due to access tracks
}
\description{
removal_at
}
