% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_Volume_of_peat_removed.R
\name{removal_bp}
\alias{removal_bp}
\title{removal_bp}
\usage{
removal_bp(pit_dims)
}
\arguments{
\item{pit_dims}{number, length, width and depth of borrow pits (list)}
}
\value{
Area/Volume of peat removed due to borrow pits
}
\description{
removal_bp
}
