% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_Volume_of_peat_removed.R
\name{removal_f}
\alias{removal_f}
\title{removal_f}
\usage{
removal_f(f_dims)
}
\arguments{
\item{f_dims}{foundation dimensions (list)}
}
\value{
Area/Volume of peat removed due to foundations
}
\description{
removal_f
}
