% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5a_Volume_of_peat_removed.R
\name{removal_h}
\alias{removal_h}
\title{removal_h}
\usage{
removal_h(h_dims)
}
\arguments{
\item{h_dims}{hardstand dimensions (list)}
}
\value{
Area/Volume of peat removed due to hardstanding
}
\description{
removal_h
}
