% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7d_Wind_speed_ratios.R
\name{Wind_speed_ratios}
\alias{Wind_speed_ratios}
\title{Wind_speed_ratios}
\usage{
Wind_speed_ratios(core.dat, forestry.dat)
}
\arguments{
\item{core.dat}{UI data}

\item{forestry.dat}{UI forestry data}
}
\value{
windspeed ratios
}
\description{
Wind_speed_ratios
}
