% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Windfarm_CO2_emission_saving.R
\name{Windfarm_CO2_emissions_saving}
\alias{Windfarm_CO2_emissions_saving}
\title{Windfarm_CO2_emissions_saving}
\usage{
Windfarm_CO2_emissions_saving()
}
\value{
total Windfarm CO2 emissions saving
}
\description{
Windfarm_CO2_emissions_saving
}
