% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Windfarm_CO2_emission_saving.R
\name{Windfarm_emissions_saving}
\alias{Windfarm_emissions_saving}
\title{Windfarm_emissions_saving}
\usage{
Windfarm_emissions_saving(e_out, E_mat)
}
\arguments{
\item{E_mat}{counterfactuals in matrix form with IDs as rownames}

\item{energy_output}{output of Windfarm_output}
}
\value{
Total Windfarm CO2 emissions savings
}
\description{
Windfarm_emissions_saving
}
