% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Windfarm_CO2_emission_saving.R
\name{Windfarm_energy_output}
\alias{Windfarm_energy_output}
\title{Windfarm_energy_output}
\usage{
Windfarm_energy_output()
}
\value{
total Windfarm energy output
}
\description{
Windfarm_energy_output
}
