% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Windfarm_CO2_emission_saving.R
\name{Windfarm_output}
\alias{Windfarm_output}
\title{Windfarm_output}
\usage{
Windfarm_output(p_cap, n_turb, c_turb, forestry.dat = NULL)
}
\arguments{
\item{p_cap}{capcity factor}

\item{n_turb}{number of turbines}

\item{c_turb}{maximum tubine power capacity}
}
\value{
Total Windfarm energy output
}
\description{
Windfarm_output
}
