% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_CO2_loss_due_to_backup.R
\name{Backup_emissions}
\alias{Backup_emissions}
\title{Backup_emissions}
\usage{
Backup_emissions(c_turb, n_turb, p_therm, p_back, E_mat, t_wf)
}
\arguments{
\item{c_turb}{maximum tubine power capacity}

\item{n_turb}{number of turbines}

\item{p_therm}{percent thermal efficiency of backup source}

\item{p_back}{percent extra capacity needed for backup}

\item{E_mat}{counterfactuals in matrix form with IDs as rownames}

\item{t_wf}{windfarm lifetime}
}
\value{
Backup lifetime emissions
}
\description{
Backup_emissions
}
