% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_CO2_loss_due_to_backup.R
\name{Backup_emissions}
\alias{Backup_emissions}
\title{Backup_emissions}
\usage{
Backup_emissions(core.dat, E_mat)
}
\arguments{
\item{core.dat}{UI data}
}
\value{
L_back
}
\description{
Backup_emissions
}
