% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_CO2_loss_by_DOC_and_POC_loss.R
\name{CO2_loss_DOC_POC}
\alias{CO2_loss_DOC_POC}
\title{CO2_loss_DOC_POC}
\usage{
CO2_loss_DOC_POC(
  core.dat,
  L_indirect,
  L_improvement,
  pC_DOC = c(Exp = 26, Min = 7, Max = 40),
  pDOC_CO2 = 100,
  pC_POC = c(Exp = 8, Min = 4, Max = 10),
  pPOC_CO2 = 100,
  CO2_C = 3.667,
  CH4_CO2 = 30.67,
  pC_CH4 = 0.75
)
}
\arguments{
\item{core.dat}{UI data}
}
\value{
L_DOC
}
\description{
CO2_loss_DOC_POC
}
