% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5b_CO2_loss_from_removed_peat.R
\name{CO2_loss_removed}
\alias{CO2_loss_removed}
\title{CO2_loss_removed}
\usage{
CO2_loss_removed(
  core.dat,
  AV_direct,
  L_indirect,
  CO2_C = 3.667,
  pCO2_lost = 100
)
}
\arguments{
\item{core.dat}{UI data}

\item{AV_direct}{Area/Volume of removed peat}

\item{L_indirect}{CO2 from drained peat}

\item{CO2_C}{Molecular weight ratio C to CO2}

\item{pCO2_lost}{percent Carbon lost as CO2}
}
\value{
Hypothetical carbon fixation of drained/removed peat
}
\description{
CO2_loss_removed
}
