% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5e_Emission_rates_from_soils.R
\name{Emissions_rates_soils}
\alias{Emissions_rates_soils}
\title{Emissions_rates_soils}
\usage{
Emissions_rates_soils(core.dat, construct.dat, AV_indirect)
}
\arguments{
\item{core.dat}{UI data}

\item{construct.dat}{UI construction data}

\item{AV_indirect}{Area/Volume of drained peat}

\item{CO2_C}{Molecular weight ratio C to CO2}

\item{CH4_CO2}{CH4 to CO2 conversion factor}
}
\value{
Emissions rates from drained and undrained peatland
}
\description{
Emissions_rates_soils
}
