% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_CO2_loss_due_to_turbine_life.R
\name{Lifetime_emissions}
\alias{Lifetime_emissions}
\title{Lifetime_emissions}
\usage{
Lifetime_emissions(core.dat, construct.dat)
}
\arguments{
\item{core.dat}{UI data}

\item{construct.dat}{UI construction data}
}
\value{
L_life
}
\description{
Lifetime_emissions
}
