% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Loss_of_CO2_fixing_pot.R
\name{Loss_of_CO2_fix_pot}
\alias{Loss_of_CO2_fix_pot}
\title{Loss_of_CO2_fix_pot}
\usage{
Loss_of_CO2_fix_pot(
  A_direct,
  A_indirect,
  G_bog,
  t_wf,
  t_restore,
  CO2_C = 3.667
)
}
\arguments{
\item{A_direct}{Area of removed peat}

\item{A_indirect}{Area of drained peat}

\item{G_bog}{Peatland carbon accumulation rate}

\item{t_wf}{Windfarm life time}

\item{t_restore}{Time to restoration after decomissioning}

\item{CO2_C}{Molecular weight ratio C to CO2}
}
\value{
Hypothetical carbon fixation of drained/removed peat
}
\description{
Loss_of_CO2_fix_pot
}
